requirements.txt
# Web框架
Django>=4.2,<5
django-plotly-dash>=2.3
djangorestframework>=3.14
django-celery-beat>=2.5
gunicorn>=21.0
whitenoise>=6.0

# 数据库
psycopg2-binary>=2.9

# 数据处理与可视化
pandas>=2.0
plotly>=5.0
dash>=2.14
dash-bootstrap-components>=1.4
numpy>=1.21

# 异步任务队列
celery>=5.3
redis>=5.0

# 网络爬虫与自动化
playwright>=1.40

# 机器学习与NLP
sentence-transformers>=2.2
scikit-learn>=1.3
torch>=2.0
transformers>=4.35
openai>=1.0

# 文件处理
openpyxl>=3.1

# 环境变量管理
python-dotenv>=1.0
django-widget-tweaks>=1.5.0
